refactor(vfs): move internal/vfs → pkg/vfs
The VFS abstraction is general-purpose and not specific to this server's
internals. Placing it in pkg/ allows future reuse by plugin tooling,
CLIs, or a plugin SDK without being gated behind internal/ visibility.

No behavior change. Nothing outside the package imports it yet.
