+++ pkg/vfs/registry.go
