+++ pkg/vfs/vfs_test.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/autobutler-org/autobutler/internal/vfs"
+	"github.com/autobutler-org/autobutler/pkg/vfs"
 )
 
 func makeLocalVFS(t *testing.T, namespaceID string) vfs.VFS {
